pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "network-loop-finder"
version = "0.1.0"
description = "Detect loops in a switch network via SSH + CDP/LLDP discovery"
requires-python = ">=3.11"
dependencies = [
    "netmiko>=4.3.0",
    "networkx>=3.3",
    "rich>=13.7.0",
    "PyYAML>=6.0",
]

[project.scripts]
loop-finder = "loop_finder.entry:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["loop_finder*"]
